+++ apps/web-antd/src/views/system/post/index.vue
@@ -133,7 +133,8 @@ function handleDownloadExcel() {
     <DeptTree
       v-model:select-dept-id="selectDeptId"
       class="w-[260px]"
-      @select="() => tableApi.query()"
+      @reload="() => tableApi.reload()"
+      @select="() => tableApi.reload()"
     />
     <BasicTable class="flex-1 overflow-hidden" table-title="岗位列表">
       <template #toolbar-tools>